+++ esp32_wireless_control/firmware/msx_motor_driver.cpp
@@ -9,7 +9,7 @@ MSxMotorDriver::MSxMotorDriver(uint8_t ms1Pin, uint8_t ms2Pin):
   pinMode(_ms2Pin, OUTPUT);
 }
 
-void MSxMotorDriver::setMicrosteps(uint8_t microsteps)
+void MSxMotorDriver::setMicrosteps(uint16_t microsteps)
 {
     switch (microsteps)
     {
